perf(analytics-engine-browser): move createFsPglite to test-utils subpath
Inlining createFsPglite into the main barrel forced downstream bundlers
to parse @electric-sql/pglite just to reach BrowserAnalyticsStore, which
tripped rolldown's resolveNewUrlToAsset pass and emitted ~14MB of
orphan pglite.wasm/pglite.data into consumer dists even after the JS
itself was tree-shaken. Expose createFsPglite via a separate
./test-utils entry so production graphs no longer reach it.
