+++ packages/analytics-engine/browser/index.ts
@@ -1,2 +1,1 @@
 export * from "./src/BrowserAnalyticsStore.js";
-export * from "./src/utils.js";
