+++ packages/analytics-engine/browser/tsdown.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "tsdown";
 
 export default defineConfig({
-  entry: ["./index.ts"],
+  entry: ["./index.ts", "./test-utils.ts"],
   outDir: "dist",
   platform: "browser",
   clean: true,
